Write files for release to separate directory
To allow us to automate release requests for this workspace, write the file
to be released (report.html) and the supporting file
(latest_import_dates.csv)
to the same containing directory. This means we can set up the release
request so that both output and supporting files are added to the same
group.